ki7/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
numpy>=1.24.0
Pillow>=10.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Optional: For production deployment
# torch>=2.0.0
# torchvision>=0.15.0
# opencv-python>=4.8.0
# scikit-learn>=1.3.0
# pandas>=2.0.0
